fix: Got sidetracked again, abstracted and reconfigured ExperienceGrid
